Fix test SegmentReplicationIndexShardTests.testPrimaryRestart.
This test is specific to remote store and should not be run for node-node replication.
